Changed Apis sizes from usize to DeviceSize for clarity

+++ src/vulkan_abstraction/acceleration_structure/mod.rs
@@ -80,7 +80,7 @@ impl AccelerationStructure {
         };
         let buffer = vulkan_abstraction::GpuOnlyBuffer::new::<u8>(
             Rc::clone(&core),
-            acceleration_structure_size_info.acceleration_structure_size as usize,
+            acceleration_structure_size_info.acceleration_structure_size,
             vk::BufferUsageFlags::ACCELERATION_STRUCTURE_STORAGE_KHR
                 | vk::BufferUsageFlags::SHADER_DEVICE_ADDRESS
                 | vk::BufferUsageFlags::STORAGE_BUFFER,
@@ -104,7 +104,7 @@ impl AccelerationStructure {
         // the scratch buffer that will be used for building the acceleration structure (and can be dropped afterwards)
         let scratch_buffer = vulkan_abstraction::GpuOnlyBuffer::new_aligned::<u8>(
             Rc::clone(&core),
-            acceleration_structure_size_info.build_scratch_size as usize,
+            acceleration_structure_size_info.build_scratch_size,
             core.device()
                 .acceleration_structure_properties()
                 .min_acceleration_structure_scratch_offset_alignment as u64,
@@ -227,7 +227,7 @@ impl AccelerationStructure {
         // the scratch buffer that will be used for building the acceleration structure (and can be dropped afterwards)
         let scratch_buffer = vulkan_abstraction::GpuOnlyBuffer::new::<u8>(
             Rc::clone(&self.core),
-            size_info.build_scratch_size as usize,
+            size_info.build_scratch_size,
             vk::BufferUsageFlags::SHADER_DEVICE_ADDRESS | vk::BufferUsageFlags::STORAGE_BUFFER,
             "acceleration structure update scratch buffer",
         )?;
